+++ Data/Primitive/PrimArray.hs
@@ -508,6 +508,8 @@ freezePrimArray !src !off !len = do
 --
 -- This operation makes a copy of the specified slice, so it is safe to
 -- use the immutable array afterward.
+--
+-- @since 0.7.2.0
 thawPrimArray
   :: (PrimMonad m, Prim a)
   => PrimArray a -- ^ source