+++ changelog.md
@@ -20,6 +20,8 @@
   * Allow building with GHC 8.12.
     Thanks Ryan GL Scott for this and every compat patch over time.
 
+  * Add `clonePrimArray` and `cloneMutablePrimArray`.
+
 ## Changes in version 0.7.0.0
 
   * Remove `Addr` data type, lifted code should use `Ptr a` now