+++ changelog.md
@@ -15,6 +15,9 @@
     `PrimArray` and `MutablePrimArray`.
     by Callan McGill
 
+  * Add `NFData` instances for `Array` and `SmallArray`.
+    by Callan McGill
+
 ## Changes in version 0.7.0.1
 
   * Allow building with GHC 8.12.